Use 32 bit quantity for packed buffer length
When debug dumping the debug registry, a ton of packed buffers
get compbined into a single packed buffer, and we would occasionally
overflow the uint16_t used to hold the length, leading to a crash in
snis_server.

+++ snis_marshal.h
@@ -17,7 +17,7 @@ struct packed_double
 struct packed_buffer
 {
 	unsigned char *buffer;
-	uint16_t buffer_size; /* size of space allocated for buffer */
+	uint32_t buffer_size; /* size of space allocated for buffer */
 	int buffer_cursor;
 };
